clean and fix some issues - [x]  warna flag rekonsiliasi grey - [x]  field form input amount di format pakai titik - [ ]  soft delete/edit upsert for short (partial done) - [x]  hutang kepada user minus nominal jadi merah - [x]  pengirim dan penerima di satu kolom wording ( rumah ke user = bayar hutang, user ke pengeluaran = rumah hutang ke user, user ke rumah bayar kontribusi) - [x]  form ketika gagal jangan tutup retain input - [x]  tunggu promise selesaii baru tutup modal - [x]  pesan input transakasi 2 detik
